requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
sqlalchemy>=2.0
asyncpg>=0.29
alembic>=1.13
pydantic>=2.6
pydantic-settings>=2.4.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
redis>=5.0
rq>=1.16
python-dotenv>=1.0
httpx>=0.27
pytest>=8
pytest-asyncio>=0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
psycopg2-binary>=2.9


